修改整个系统的权限系统为部门级别权限认证

+++ sheepyu-module-system/sheepyu-module-system-biz/src/main/java/top/sheepyu/module/system/controller/admin/user/vo/SystemUserQueryVo.java
@@ -31,7 +31,7 @@ public class SystemUserQueryVo extends PageParam {
     private Integer type;
 
     @ApiModelProperty("部门id")
-    private Integer deptId;
+    private Long deptId;
 
     @ApiModelProperty("最近登录时间")
     @DateTimeFormat(pattern = DATE_TIME_FORMAT)
